+++ releases/ChangeLog.md
@@ -5,6 +5,7 @@
 - Enhancement: Allow to specify start index for address explorer export and browsing
 - Enhancement: Allow unlimited index for BIP-85 derivations. Needs to be enabled first in `Danger Zone` 
 - Enhancement: Add `Nunchuk` option to `Export Wallet`
+- Enhancement: `View Identity` shows temporary seed active on the top 
 - Change: `Passphrase` menu item is no longer offered if BIP39 passphrase
   already in use. Use `Restore Master` with ability to keep or purge current
   passphrase wallet settings.